requirements.txt
easydict==1.9
Flask==1.1.2
numpy==1.19.2
opencv_python==4.5.5.64
pandas==1.3.4
Pillow==9.1.1
tensorflow-cpu==2.4.1
